Created ProjectRepository. Changed relevant methods in TaskCompassController and UserRepository because of addition of ProjectRepository
