+++ src/main/java/com/alphaS/alphasolutions/controllers/TaskCompassController.java
@@ -3,6 +3,7 @@ package com.alphaS.alphasolutions.controllers;
 
 
 import com.alphaS.alphasolutions.model.UserModel;
+import com.alphaS.alphasolutions.repositories.ProjectRepository;
 import com.alphaS.alphasolutions.repositories.UserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
@@ -22,10 +23,12 @@ import java.util.List;
 @Controller
 public class TaskCompassController {
 
-    private final UserRepository repository;
+    private final UserRepository userRepository;
+    private final ProjectRepository projectRepository;
 
-    public TaskCompassController(UserRepository repository) {
-        this.repository = repository;
+    public TaskCompassController(UserRepository userRepository, ProjectRepository projectRepository) {
+        this.userRepository = userRepository;
+        this.projectRepository = projectRepository;
     }
 
     @GetMapping("")
@@ -40,7 +43,7 @@ public class TaskCompassController {
     @PostMapping("/signin")
     public String signinpostmapping(@RequestParam("username") String username, @RequestParam("password") String password, HttpSession session, Model model){
         try {
-            UserModel usermodel = repository.signin(username, password);
+            UserModel usermodel = userRepository.signin(username, password);
             session.setAttribute("user", usermodel); // store user in session
             return "dashboard"; // return dashboard page if login is successful
         } catch (SQLException e) {
@@ -65,7 +68,7 @@ public class TaskCompassController {
                                                @RequestParam String country,
                                                @RequestParam String clientId) {
         try {
-            String result = repository.createClient(clientName, contactPoNo, contactPerson, companyPoNo, address, zipCode, country, clientId);
+            String result = projectRepository.createClient(clientName, contactPoNo, contactPerson, companyPoNo, address, zipCode, country, clientId);
             return ResponseEntity.ok(result);
         } catch (SQLException e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to create client");
@@ -82,7 +85,7 @@ public class TaskCompassController {
                                              @RequestParam String country,
                                              @RequestParam String clientId) {
         try {
-            String result = repository.editClient(clientName, contactPoNo, contactPerson, companyPoNo, address, zipCode, country, clientId);
+            String result = projectRepository.editClient(clientName, contactPoNo, contactPerson, companyPoNo, address, zipCode, country, clientId);
             return ResponseEntity.ok(result);
             } catch (SQLException e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to edit client");
@@ -94,7 +97,7 @@ public class TaskCompassController {
     @ResponseBody
     public ResponseEntity<String> searchClient(@RequestParam String clientName) {
         try {
-            String result = repository.searchClients(clientName).toString();
+            String result = projectRepository.searchClients(clientName).toString();
             return ResponseEntity.ok(result);
         } catch (SQLException e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to search client");
@@ -103,7 +106,7 @@ public class TaskCompassController {
 
     @PostMapping("/clients/delete")
     public String deleteClient(@RequestParam int clientID){
-        String message = repository.deleteClient(clientID);
+        String message = projectRepository.deleteClient(clientID);
         return message;
     }
 
@@ -112,7 +115,7 @@ public class TaskCompassController {
     @ResponseBody
     public ResponseEntity<String> searchProject(@RequestParam String projectName) {
         try {
-            String result = repository.searchProjects(projectName).toString();
+            String result = projectRepository.searchProjects(projectName).toString();
             return ResponseEntity.ok(result);
         } catch (SQLException e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to search project");
@@ -120,7 +123,7 @@ public class TaskCompassController {
     }
     @PostMapping("project/delete")
     public String deleteProject(@RequestParam int projectID){
-        String message = repository.deleteProject(projectID);
+        String message = projectRepository.deleteProject(projectID);
         return message;
     }
 
@@ -129,7 +132,7 @@ public class TaskCompassController {
     public ResponseEntity<String> createTeam(HttpServletRequest request) {
         try {
             String teamName = request.getParameter("teamName");
-            int teamId = repository.createTeam(teamName);
+            int teamId = projectRepository.createTeam(teamName);
             return ResponseEntity.ok().body(Integer.toString(teamId));
         } catch (SQLException e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to create team");
@@ -143,7 +146,7 @@ public class TaskCompassController {
             int teamId = Integer.parseInt(request.getParameter("teamId"));
             String[] usernames = request.getParameterValues("usernames[]");
             List<String> userNames = Arrays.asList(usernames);
-            repository.addTeamMembers(teamId, userNames);
+            projectRepository.addTeamMembers(teamId, userNames);
             return ResponseEntity.ok().body("Team members added successfully.");
         } catch (SQLException e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to add team members.");
